added a check to the deploy script (protects agains accidentally deploying to a stage that doesn't exist - the stage will be created)

+++ cloud-logic/lambda/deploy.py
@@ -208,6 +208,9 @@ if __name__ == "__main__":
     source_parent = './src'
     base_yaml_file = './serverless-base.yml'
     stage = input('stage: ')
+    if not stage in ('dev', 'prod'):
+        print('Invalid stage')
+        exit(1)
     run_build(handler_implementations_parent_dir, target_build_directory, source_parent, base_yaml_file)
     os.system(f"cd build; serverless deploy --stage {stage} --force")
     shutil.rmtree(target_build_directory)
